Add unit tests for the Kernel
Also refactor Manager to two classes => Worker, Lister. Manager didn't
really say anything about its purpose and would potentially end up
doing a lot of unrelated things.
In the Config class the driver now is a singleton, but it's reset when
the config path is changed.
Also, turning the queues from the config into an array of Queue objects
is moved to the Config class as a static method.