Add tests for Worker
Also improve checking for running processes in environment where the
ps command isn't available. And further introduce a helper method that
tries a callback until it's returning something else than false (or max
tries is reached), which is better than waiting for a fixed timespan for
something to be done, because that can vary in different environments
and therefore you'd always have to choose a pretty high sleep time.