Improve handling processes
The worker process now writes its pid and the current time to a file in
the datapath. It updates the time on every loop iteration (default
0.2 seconds). When checking if a worker process is already working the
queues, it checks if there is a pid and time in that file, if the pid
is still running and if the time is not older than one second. So, if
a worker process dies but becomes a zombie, the library will know that
it isn't working the queue anymore at max a second later.

Moved the functionality to get information about running processes to
the Processes class. The Process class now only has the functionality
for the one process it represents.