Fix writing files
There was an issue when writing to files when using stream resource and
ftruncate() isn't called before fwrite(). If the new content is shorter
than the content before, it writes the new content from position 0 (as
it's always rewinding after reading), but the part from the previous
content that is longer than the new content remains at the end of the
file.

For example, if this is the content before writing:
`old content that is longer than the new content`

And this is the new content you want to write:
`new shorter content`

Actually the content after write will be:
`new shorter content longer than the new content`

Weirdly unserialize seems to still work with such more or less broken
file content.