Add event listeners feature
Via the config you can add listeners (classes implementing the
QueueEventListener interface) to certain queues. The available events
are: waiting, running, finished, failed, lost, cancelled. When a
listener is added to a queue in the config it will be called with each
job that was just added (waiting), started running (running), just
finished, and so on.