Fix issue with writing the index file
When the new content of the ppq index file was shorter than the old
content, it wrote the new content starting from the beginning of the
file and the part of the old content that was longer than the new
content, remained at the end of the file. Fixed it be truncating the
file before writing the new content.